Refactor readFunctions to return features
Instead of storing them on the module. This required factoring LEB
code into its own header.
