+++ src/wasm-binary.h
@@ -29,8 +29,9 @@
 #include "wasm-traversal.h"
 #include "asmjs/shared-constants.h"
 #include "asm_v_wasm.h"
-#include "wasm-builder.h"
+#include "leb.h"
 #include "parsing.h"
+#include "wasm-builder.h"
 #include "wasm-validator.h"
 #include "ir/import-utils.h"
 
@@ -49,99 +50,6 @@ enum WebLimitations {
   MaxFunctionLocals = 50 * 1000
 };
 
-template<typename T, typename MiniT>
-struct LEB {
-  static_assert(sizeof(MiniT) == 1, "MiniT must be a byte");
-
-  T value;
-
-  LEB() = default;
-  LEB(T value) : value(value) {}
-
-  bool hasMore(T temp, MiniT byte) {
-    // for signed, we must ensure the last bit has the right sign, as it will zero extend
-    return std::is_signed<T>::value ? (temp != 0 && temp != T(-1)) || (value >= 0 && (byte & 64)) || (value < 0 && !(byte & 64)) : (temp != 0);
-  }
-
-  void write(std::vector<uint8_t>* out) {
-    T temp = value;
-    bool more;
-    do {
-      uint8_t byte = temp & 127;
-      temp >>= 7;
-      more = hasMore(temp, byte);
-      if (more) {
-        byte = byte | 128;
-      }
-      out->push_back(byte);
-    } while (more);
-  }
-
-  // @minimum: a minimum number of bytes to write, padding as necessary
-  // returns the number of bytes written
-  size_t writeAt(std::vector<uint8_t>* out, size_t at, size_t minimum = 0) {
-    T temp = value;
-    size_t offset = 0;
-    bool more;
-    do {
-      uint8_t byte = temp & 127;
-      temp >>= 7;
-      more = hasMore(temp, byte) || offset + 1 < minimum;
-      if (more) {
-        byte = byte | 128;
-      }
-      (*out)[at + offset] = byte;
-      offset++;
-    } while (more);
-    return offset;
-  }
-
-  LEB<T, MiniT>& read(std::function<MiniT()> get) {
-    value = 0;
-    T shift = 0;
-    MiniT byte;
-    while (1) {
-      byte = get();
-      bool last = !(byte & 128);
-      T payload = byte & 127;
-      typedef typename std::make_unsigned<T>::type mask_type;
-      auto shift_mask = 0 == shift
-                            ? ~mask_type(0)
-                            : ((mask_type(1) << (sizeof(T) * 8 - shift)) - 1u);
-      T significant_payload = payload & shift_mask;
-      if (significant_payload != payload) {
-        if (!(std::is_signed<T>::value && last)) {
-          throw ParseException("LEB dropped bits only valid for signed LEB");
-        }
-      }
-      value |= significant_payload << shift;
-      if (last) break;
-      shift += 7;
-      if (size_t(shift) >= sizeof(T) * 8) {
-        throw ParseException("LEB overflow");
-      }
-    }
-    // If signed LEB, then we might need to sign-extend. (compile should
-    // optimize this out if not needed).
-    if (std::is_signed<T>::value) {
-      shift += 7;
-      if ((byte & 64) && size_t(shift) < 8 * sizeof(T)) {
-        size_t sext_bits = 8 * sizeof(T) - size_t(shift);
-        value <<= sext_bits;
-        value >>= sext_bits;
-        if (value >= 0) {
-          throw ParseException(" LEBsign-extend should produce a negative value");
-        }
-      }
-    }
-    return *this;
-  }
-};
-
-typedef LEB<uint32_t, uint8_t> U32LEB;
-typedef LEB<uint64_t, uint8_t> U64LEB;
-typedef LEB<int32_t, int8_t> S32LEB;
-typedef LEB<int64_t, int8_t> S64LEB;
 
 //
 // We mostly stream into a buffer as we create the binary format, however,
@@ -1080,7 +988,6 @@ public:
   void readFunctionTableDeclaration();
   void readTableElements();
   void readNames(size_t);
-  void readFeatures(const UserSection&);
 
   // Debug information reading helpers
   void setDebugLocations(std::istream* sourceMap_) {
